+++ pkg/config/measurementcfg.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+
+	"github.com/toni-moreno/snmpcollector/pkg/data/utils"
 )
 
 //MeasurementCfg the measurement configuration
@@ -26,9 +28,50 @@ type MeasurementCfg struct {
 	Description   string           `xorm:"description"`
 }
 
-//CheckComputedMetric check for computed metrics
-func (mc *MeasurementCfg) CheckComputedMetric() error {
-	parameters := make(map[string]interface{})
+//CheckComputedMetric2 check for computed metrics based on check if variable definition exist
+func (mc *MeasurementCfg) CheckComputedMetric2(parameters map[string]interface{}) error {
+	var extvars []string
+	var allusablevars []string
+
+	//external defined vars ( from Catalog database)
+	for k := range parameters {
+		extvars = append(extvars, k)
+	}
+	allusablevars = append(allusablevars, extvars...)
+	//internall defined vars( FieldNames and FR/NR)
+	intvars, _ := mc.GetInternalVarNames()
+	allusablevars = append(allusablevars, intvars...)
+
+	for _, val := range mc.EvalMetric {
+		varinmetric, err := val.GetUsedVarNames()
+		if err != nil {
+			return fmt.Errorf("Error on Get Variables on Metric %s ERROR: %s", val.ID, err)
+		}
+		log.Debugf("checking if existing var in measurement: %s Metric: %s", mc.ID, val.ID)
+
+		for _, varin := range varinmetric {
+			//check if exist on usable vars
+			found := false
+			for _, varusable := range allusablevars {
+				if varin == varusable {
+					found = true
+					break
+				}
+			}
+			log.Debugf("variable %s : FOUND: %t ", varin, found)
+			if !found {
+				return fmt.Errorf("variable %s defined in Metric %s not Found  (in any of other metric FieldNames of the measurement, NF, RF and Variable Catalog) ", varin, val.ID)
+			}
+		}
+
+	}
+
+	return nil
+}
+
+/*/CheckComputedMetric check for computed metrics based on  Evalutation Execution
+func (mc *MeasurementCfg) CheckComputedMetric(parameters map[string]interface{}) error {
+
 	log.Debugf("Building check parrameters array for index measurement %s", mc.ID)
 	parameters["NR"] = 1                   //Number of rows (like awk)
 	parameters["NF"] = len(mc.FieldMetric) //Number of fields ( like awk)
@@ -38,18 +81,83 @@ func (mc *MeasurementCfg) CheckComputedMetric() error {
 	}
 	log.Debugf("PARAMETERS: %+v", parameters)
 	//compute Evalutated metrics
+	var err error
+	var errstr []string
 	for _, v := range mc.EvalMetric {
-		err := v.CheckEvalCfg(parameters)
+		err = v.CheckEvalCfg(parameters)
 		if err != nil {
-			return fmt.Errorf("Error on metric %s evaluation ERROR : %s", v.ID, err)
+			str := fmt.Sprintf("Error on metric %s evaluation ERROR : %s", v.ID, err)
+			log.Warnf("%s", str)
+			errstr = append(errstr, str)
+			//return fmt.Errorf("Error on metric %s evaluation ERROR : %s", v.ID, err)
 		}
 		parameters[v.FieldName] = float64(1)
 	}
+	if err != nil {
+		return fmt.Errorf("%s", strings.Join(errstr, "\n"))
+	}
 	return nil
+}*/
+
+// GetInternalVarNames returns an string array with all posible internal variable Names
+func (mc *MeasurementCfg) GetInternalVarNames() ([]string, error) {
+	var intvars []string // internal ( from metric ID's)
+	//Get InternalVariables (those from field names and "NF" and "NR")
+	for _, val := range mc.FieldMetric {
+		intvars = append(intvars, val.FieldName)
+	}
+	for _, val := range mc.EvalMetric {
+		intvars = append(intvars, val.FieldName)
+	}
+	for _, val := range mc.OidCondMetric {
+		intvars = append(intvars, val.FieldName)
+	}
+	intvars = append(intvars, "NF")
+	intvars = append(intvars, "NR")
+	//log.Debugf("INTVARS %s : %#+v ", mc.ID, intvars)
+	return utils.RemoveDuplicatesUnordered(intvars), nil
+}
+
+// GetAllUsedVarNamesInMetrics return all used Var names in Metrics
+func (mc *MeasurementCfg) GetAllUsedVarNamesInMetrics() ([]string, error) {
+	var allvars []string // all
+	// Get All used Variables in Formulas
+	for _, val := range mc.EvalMetric {
+		vars, err := val.GetUsedVarNames()
+		if err != nil {
+			return nil, err
+		}
+		log.Debugf("checking if existing var sin measurment %s : %#+v ", mc.ID, vars)
+		allvars = append(allvars, vars...)
+	}
+
+	log.Debugf("ALLVARS %s : %#+v ", mc.ID, allvars)
+	return utils.RemoveDuplicatesUnordered(allvars), nil
+}
+
+// GetExternalVars Get Needed External Variables in this Measurement
+func (mc *MeasurementCfg) GetExternalVars() ([]string, error) {
+
+	intvars, err := mc.GetInternalVarNames()
+	if err != nil {
+		return nil, err
+	}
+
+	allvars, err := mc.GetAllUsedVarNamesInMetrics()
+	if err != nil {
+		return nil, err
+	}
+
+	//get difference
+	extvars := utils.DiffSlice(intvars, allvars)
+
+	log.Debugf("EXTVARS %s : %#+v ", mc.ID, extvars)
+
+	return extvars, nil
 }
 
 //Init initialize the measurement configuration
-func (mc *MeasurementCfg) Init(MetricCfg *map[string]*SnmpMetricCfg) error {
+func (mc *MeasurementCfg) Init(MetricCfg *map[string]*SnmpMetricCfg, varmap map[string]interface{}) error {
 	//mc.ID = name
 	//validate config values
 	if len(mc.Name) == 0 {
@@ -151,8 +259,12 @@ func (mc *MeasurementCfg) Init(MetricCfg *map[string]*SnmpMetricCfg) error {
 		fieldnamecheckarray[v.FieldName] = v.ID
 	}
 	//Check if all evaluated metrics has well defined its parameters as FieldNames
-	err := mc.CheckComputedMetric()
+	if varmap == nil {
+		varmap = make(map[string]interface{})
+	}
+	err := mc.CheckComputedMetric2(varmap)
 	if err != nil {
+		log.Warnf(" This computed Metric has some errors!! please review its definition", err)
 		return err
 	}