+++ pkg/data/impexp/import.go
@@ -149,6 +149,21 @@ func (e *ExportData) ImportCheck() (*ExportData, error) {
 				o.Error = fmt.Sprintf("Duplicated object %s in the database", o.ObjectID)
 				duplicated = append(duplicated, o)
 			}
+		case "varcatalogcfg":
+			data := config.VarCatalogCfg{}
+			json.Unmarshal(raw, &data)
+			ers := binding.RawValidate(data)
+			if ers.Len() > 0 {
+				e, _ := json.Marshal(ers)
+				o.Error = string(e)
+				duplicated = append(duplicated, o)
+				break
+			}
+			_, err := dbc.GetVarCatalogCfgByID(o.ObjectID)
+			if err == nil {
+				o.Error = fmt.Sprintf("Duplicated object %s in the database", o.ObjectID)
+				duplicated = append(duplicated, o)
+			}
 		default:
 			return &ExportData{Info: e.Info, Objects: duplicated}, fmt.Errorf("Unknown type object type %s ", o.ObjectTypeID)
 		}
@@ -360,6 +375,29 @@ func (e *ExportData) Import(overwrite bool, autorename bool) error {
 			if err != nil {
 				return err
 			}
+		case "varcatalogcfg":
+			log.Debugf("Importing varcatalogcfg : %+v", o.ObjectCfg)
+			data := config.VarCatalogCfg{}
+			json.Unmarshal(raw, &data)
+			var err error
+			_, err = dbc.GetVarCatalogCfgByID(o.ObjectID)
+			if err == nil { //value exist already in the database
+				if overwrite == true {
+					_, err2 := dbc.UpdateVarCatalogCfg(o.ObjectID, data)
+					if err2 != nil {
+						return fmt.Errorf("Error on overwrite object [%s] %s : %s", o.ObjectTypeID, o.ObjectID, err2)
+					}
+					break
+				}
+			}
+			if autorename == true {
+				data.ID = data.ID + suffix
+			}
+			_, err = dbc.AddVarCatalogCfg(data)
+			if err != nil {
+				return err
+			}
+
 		default:
 			return fmt.Errorf("Unknown type object type %s ", o.ObjectTypeID)
 		}